move stores to store.ts
